updates to ctr

+++ src/Controller/NotificationsController.php
@@ -103,14 +103,24 @@ class NotificationsController extends AppController
      */
     public function delete($id = null)
     {
+        //TODO: Use session AUTH to ensure the user is assigned notification before soft delete
         $this->request->allowMethod(['post', 'delete']);
-        $notification = $this->Notifications->get($id);
-        if ($this->Notifications->delete($notification)) {
-            $this->Flash->success(__('The notification has been deleted.'));
-        } else {
-            $this->Flash->error(__('The notification could not be deleted. Please, try again.'));
-        }
+        if($this->request->is('json')) {
+            $notification = $this->Notifications->get($this->request->data['id']);
+            if ($this->Notifications->delete($notification)) {
+                $this->set([
+                        'message' => 'Notification successfully deleted', 
+                        '_serialize' => ['message']]);
+                return;
+            } else {
+                $this->response->body('Failure');
+                $this->response->statusCode(403);
+                $this->set([
+                    'message' => 'Unable to delete!!', 
+                    '_serialize' => ['message']]);
+                return;
+            }
 
-        return $this->redirect(['action' => 'index']);
+        }
     }
 }
